:sparkles: Add Duration variants for ClientHttpRequestFactory timeouts
This commit adds overloaded variants of timeout setters that take a
Duration instead of a number.

+++ today-web/src/main/java/cn/taketoday/http/client/JettyClientHttpRequest.java
@@ -1,8 +1,5 @@
 /*
- * Original Author -> Harry Yang (taketoday@foxmail.com) https://taketoday.cn
- * Copyright © Harry Yang & 2017 - 2023 All Rights Reserved.
- *
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER
+ * Copyright 2017 - 2023 the original author or authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -53,9 +50,9 @@ class JettyClientHttpRequest extends AbstractStreamingClientHttpRequest {
 
   private final Request request;
 
-  private final int readTimeout;
+  private final long readTimeout;
 
-  public JettyClientHttpRequest(Request request, int readTimeout) {
+  public JettyClientHttpRequest(Request request, long readTimeout) {
     this.request = request;
     this.readTimeout = readTimeout;
   }
